+++ backend/src/reservations/reservations.module.ts
@@ -3,9 +3,13 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { Reservation } from '../models/reservation.entity';
 import { ReservationsController } from './reservations.controller';
 import { ReservationsService } from './reservations.service';
+import { NotificationsModule } from '../notifications/notifications.module';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([Reservation])],
+  imports: [
+    TypeOrmModule.forFeature([Reservation]),
+    NotificationsModule,
+  ],
   controllers: [ReservationsController],
   providers: [ReservationsService],
 })